Better error messages
This should provide somewhat better error messages. In case we actually
reach GitHub (and not fail because of a network error), user will see
the HTTP error code and contents of message field from the response.
It's not perfect - ideally, we should use GitHub's response to show
specific error messages (like 'duplicate filename), but that's a
task for another day.